Upload a html document to RSpace

This function can upload a html document (e.g., generated from quarto) to an RSpace Basic Document, or to a Structured Document if the template is also provided.

Usage

document_create_from_html(
  path,
  template_id = NULL,
  folder_id = NULL,
  tags = NULL,
  attachment = NULL,
  existing_document_id = NULL,
  api_key = get_api_key()
)

Arguments

path

html document to upload

template_id

document id of the RSpace template used. Overwritten by the template of existing_document_id if that one is specified.

folder_id

folder_id in which the document will be created (can be a notebook)

tags

vector of tags to apply to the document

attachment

attachment to attach to one of the fields, e.g., list(field = 7, path = "file.txt")

existing_document_id

if you want to replace a document by a new one, specify the current identifier here.

api_key

RSpace API key